requirements-dev.txt
coverage==7.6.1
djlint==1.35.2
pytest-playwright==0.5.2
ruff==0.6.8

-r requirements.txt